+++ TLM/TLM/Patch/_LoadAssetPanel/OnLoadPatch.cs
@@ -1,0 +1,29 @@
+namespace TrafficManager.Patch._LoadAssetPanel {
+    using ColossalFramework.UI;
+    using HarmonyLib;
+    using TrafficManager.State.Asset;
+
+    [HarmonyPatch(typeof(LoadAssetPanel), "OnLoad")]
+    public static class OnLoadPatch {
+        /// <summary>
+        /// when loading asset from a file, IAssetData.OnAssetLoaded() is called for all assets but the one that is loaded from the file.
+        /// this postfix calls IAssetData.OnAssetLoaded() for asset loaded from file.
+        /// </summary>
+        public static void Postfix(LoadAssetPanel __instance, UIListBox ___m_SaveList) {
+            // Taken from LoadAssetPanel.OnLoad
+            var selectedIndex = ___m_SaveList.selectedIndex;
+            var getListingMetaDataMethod = typeof(LoadSavePanelBase<CustomAssetMetaData>).GetMethod(
+                "GetListingMetaData", System.Reflection.BindingFlags.Instance | System.Reflection.BindingFlags.NonPublic);
+            var listingMetaData = (CustomAssetMetaData)getListingMetaDataMethod.Invoke(__instance, new object[] { selectedIndex });
+
+            // Taken from LoadingManager.LoadCustomContent
+            if (listingMetaData.userDataRef != null) {
+                AssetDataWrapper.UserAssetData userAssetData = listingMetaData.userDataRef.Instantiate() as AssetDataWrapper.UserAssetData;
+                if (userAssetData == null) {
+                    userAssetData = new AssetDataWrapper.UserAssetData();
+                }
+                AssetDataExtension.Instance.OnAssetLoaded(listingMetaData.name, ToolsModifierControl.toolController.m_editPrefabInfo, userAssetData.Data);
+            }
+        }
+    }
+}
