+++ TLM/TLM/TLM.csproj
@@ -166,9 +166,12 @@
     <Compile Include="Patch\_CitizenManager\ReleaseCitizenPatch.cs" />
     <Compile Include="Patch\_DefaultTool\OnToolGUIPatch.cs" />
     <Compile Include="Patch\_InfoManager\SetModePatch.cs" />
+    <Compile Include="Patch\_LoadAssetPanel\OnLoadPatch.cs" />
     <Compile Include="Patch\_NetSegment\CalculateSegmentPatch.cs" />
     <Compile Include="Patch\_RoadWorldInfoPanel\OnAdjustRoadButtonPatch.cs" />
-    <Compile Include="State\AssetDataExtension.cs" />
+    <Compile Include="State\Asset\AssetData.cs" />
+    <Compile Include="State\Asset\AssetDataExtension.cs" />
+    <Compile Include="State\SerializationUtil.cs" />
     <Compile Include="Traffic\LegacyExtVehicleType.cs" />
     <Compile Include="UI\Helpers\NodeLaneMarker.cs" />
     <Compile Include="UI\Helpers\Prompt.cs" />
