+++ TLM/TLM/Util/Record/RecordUtil.cs
@@ -1,29 +1,12 @@
-using System;
-using System.IO;
-using System.Runtime.Serialization.Formatters;
-using System.Runtime.Serialization.Formatters.Binary;
-
 namespace TrafficManager.Util.Record {
+    using TrafficManager.State;
+
+    // This API is used by MoveIT mod.
     public static class RecordUtil {
-        static BinaryFormatter GetBinaryFormatter =>
-            new BinaryFormatter { AssemblyFormat = FormatterAssemblyStyle.Simple };
+        public static IRecordable Deserialize(byte[] data) =>
+            SerializationUtil.Deserialize(data) as IRecordable;
 
-        public static IRecordable Deserialize(byte[] data) {
-            if (data == null)
-                return null;
-            //Log.Debug($"SerializationUtil.Deserialize(data): data.Length={data?.Length}");
-
-            var memoryStream = new MemoryStream();
-            memoryStream.Write(data, 0, data.Length);
-            memoryStream.Position = 0;
-            return GetBinaryFormatter.Deserialize(memoryStream) as IRecordable;
-        }
-
-        public static byte[] Serialize(IRecordable obj) {
-            var memoryStream = new MemoryStream();
-            GetBinaryFormatter.Serialize(memoryStream, obj);
-            memoryStream.Position = 0; // redundant
-            return memoryStream.ToArray();
-        }
+        public static byte[] Serialize(IRecordable obj) =>
+            SerializationUtil.Serialize(obj);
     }
 }
